Stream_Processing/Consumer/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-bigquery==2.30.1
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
kafka-python==2.0.2
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
requests==2.26.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
